+++ example_2/lib/example.dart
@@ -1,1 +1,2 @@
-//pana
+// visit https://github.com/4uzhoy/string_collector/tree/main/example or
+// https://github.com/4uzhoy/string_collector/tree/main/example_2 to see how to use example