+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## 1.0.0
-
+- Add default config
 - Initial version.